rag-app/pyproject.toml
[tool.poetry]
name = "rag-app"
version = "0.1.0"
description = "A simple rag application to help AI engineers in training on Oxford's GenAI, Cloud and MLOps course."
authors = ["Andrew P. McMahon"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sentence-transformers = "^3.1.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
